rename Go package "deploy" -> "deploymetadata"
This frees up the "deploy" package to group anything for actually related to pushing and loading.
