+++ img_tool/cmd/deploymetadata/merge.go
@@ -1,4 +1,4 @@
-package deploy
+package deploymetadata
 
 import (
 	"context"
