+++ img_tool/cmd/deploymetadata/metadata.go
@@ -1,4 +1,4 @@
-package deploy
+package deploymetadata
 
 import (
 	"bytes"
