+++ img_tool/cmd/img/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/compress",
-        "//cmd/deploy",
+        "//cmd/deploymetadata",
         "//cmd/dockersave",
         "//cmd/downloadblob",
         "//cmd/expandtemplate",
